% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_torch.R
\name{bundle.luz_module_fitted}
\alias{bundle.luz_module_fitted}
\title{Torch bundlers}
\usage{
\method{bundle}{luz_module_fitted}(x, ...)
}
\description{
For now, bundling methods for torch are only available
via the luz package, "a higher level API for torch providing
abstractions to allow for much less verbose training loops."
}
\details{
These bundlers rely on serialization methods from luz and torch,
which are described by the package authors as "experimental"
and not for "use for long term storage."
}
