% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\alias{unbundle}
\title{Bundling}
\usage{
bundle(x, ...)

unbundle(x)
}
\arguments{
\item{x}{A model object to bundle.}

\item{...}{Additional arguments to bundle methods.}
}
\description{
\code{bundle()} methods provide a consistent interface to serialization
methods for statistical model objects. The outputted bundle can be saved,
re-loaded into a new R session, and \code{unbundle()}d in a new R session for
use in prediction.
}
\details{
To read more about bundling and serialization, see
\code{vignette("bundle")}.

To see a list of currently available bundlers, see
\code{methods(bundle)}.
}
