% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bundle_description}
\alias{bundle_description}
\alias{_PACKAGE}
\alias{bundle-package}
\title{bundle: Serialize Objects With A Consistent Interface}
\description{
R holds most objects in memory. However, some models store their data in locations that are not included when one uses `save()` or `saveRDS()`. bundle provides a common API to capture this information, situate it within a portable object, and restore it for use in new settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/bundle}
  \item \url{https://rstudio.github.io/bundle/}
  \item Report bugs at \url{https://github.com/rstudio/bundle/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item Simon Couch \email{simonpatrickcouch@gmail.com}
  \item Qiushi Yan \email{qiushi.yann@gmail.com}
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
