% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_model_fit.R
\name{bundle.model_fit}
\alias{bundle.model_fit}
\title{Bundle a parsnip \code{model_fit} object}
\usage{
\method{bundle}{model_fit}(x, ...)
}
\arguments{
\item{x}{A \link[parsnip:model_fit]{model_fit} object returned
from \link[parsnip:parsnip-package]{parsnip} or other tidymodels packages.}

\item{...}{Not used in this bundler and included for compatibility with
the generic only. Additional arguments passed to this method will return
an error.}
}
\value{
A bundle object with subclass \code{bundled_model_fit}.

Bundles are a list subclass with two components:

\item{object}{An R object. Gives the output of native serialization
methods from the model-supplying package, sometimes with additional
classes or attributes that aid portability. This is often
a \link[base:raw]{raw} object.}
\item{situate}{A function. The \code{situate()} function is defined when
\code{\link[=bundle]{bundle()}} is called, though is a loose analogue of an \code{\link[=unbundle]{unbundle()}} S3
method for that object. Since the function is defined on \code{\link[=bundle]{bundle()}}, it
has access to references and dependency information that can
be saved alongside the \code{object} component. This allows for more
resilient serialization.}

Bundles are R objects that represent a "standalone" version of their
analogous model object. Thus, bundles are ready for saving to file---saving
with \code{\link[base:readRDS]{base::saveRDS()}} is our recommended serialization strategy for bundles,
unless documented otherwise for a specific method.

To restore the inputted model object \code{x} in a new environment, load its
bundle with \code{\link[base:readRDS]{base::readRDS()}} and run \code{\link[=unbundle]{unbundle()}} on it. The output
of \code{\link[=unbundle]{unbundle()}} is a model object that is ready to \code{\link[=predict]{predict()}} on new data,
and other restored functionality (like plotting or summarizing) is supported
as a side effect only.

See \code{vignette("bundle")} for more information on bundling and its motivation.
}
\description{
Bundling a model prepares it to be saved to file and later
restored for prediction in a new R session. See the 'Value' section for
more information on bundles and their usage.
}
\details{
Primarily, these tidiers call \code{\link[=bundle]{bundle()}} on the output of
\code{\link[parsnip:reexports]{parsnip::extract_fit_engine()}}. See the class of the output of that
function for more details on the bundling method for that object.
}
