% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_package.R
\docType{package}
\name{gift_description}
\alias{gift_description}
\alias{_PACKAGE}
\alias{gift_package}
\alias{gift}
\title{gift: Serialize Objects With A Consistent Interface}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Simon Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Qiushi Yan \email{qiushi.yann@gmail.com}
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
