% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\alias{unbundle}
\title{Bundling}
\usage{
bundle(x)

unbundle(x)
}
\arguments{
\item{x}{An R object to bundle.}
}
\description{
\code{bundle()} methods provide a consistent interface to serialization
methods. The outputted bundle can be \code{unbundle()}d in a new R session
and used as desired.
}
\details{
To read more about bundling and serialization, see
\code{?vignette("bundle", package = "bundle")}!
}
