% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gift_package.R
\docType{package}
\name{bundle_description}
\alias{bundle_description}
\alias{_PACKAGE}
\alias{bundle_package}
\alias{bundle}
\title{bundle: Serialize Objects With A Consistent Interface}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/bundle}
  \item Report bugs at \url{https://github.com/simonpcouch/bundle/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Qiushi Yan \email{qiushi.yann@gmail.com}
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
