% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bundle_description}
\alias{bundle_description}
\alias{_PACKAGE}
\alias{bundle-package}
\title{bundle: Serialize Objects With A Consistent Interface}
\description{
R holds most objects in memory. However, some models store their data in locations that are not included when one uses `save()` or `saveRDS()`. bundle provides a common API to capture this information, situate it within a portable object, and restore it for use in new settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/bundle}
  \item Report bugs at \url{https://github.com/simonpcouch/bundle/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Qiushi Yan \email{qiushi.yann@gmail.com}
  \item Max Kuhn \email{max@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
