% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{bundle_description}
\alias{bundle_description}
\alias{bundle-package}
\title{bundle: Serialize Model Objects With A Consistent Interface}
\description{
Typically, models in 'R' exist in memory and can be saved via regular 'R' serialization. However, some models store information in locations that cannot be saved using 'R' serialization alone. The goal of 'bundle' is to provide a common interface to capture this information, situate it within a portable object, and restore it for use in new settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/bundle}
  \item \url{https://rstudio.github.io/bundle/}
  \item Report bugs at \url{https://github.com/rstudio/bundle/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@posit.co} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item Simon Couch \email{simonpatrickcouch@gmail.com}
  \item Qiushi Yan \email{qiushi.yann@gmail.com}
  \item Max Kuhn \email{max@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
