% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_lightgbm.R
\name{bundle.lgb.Booster}
\alias{bundle.lgb.Booster}
\title{Bundle an \code{lgb.Booster} object}
\usage{
\method{bundle}{lgb.Booster}(x, ...)
}
\arguments{
\item{x}{An \code{lgb.Booster} object returned from \code{\link[lightgbm:lgb.train]{lightgbm::lgb.train()}}.}

\item{...}{Not used in this bundler and included for compatibility with
the generic only. Additional arguments passed to this method will return
an error.}
}
\value{
A bundle object with subclass \code{bundled_lgb.Booster}.

Bundles are a list subclass with two components:

\item{object}{An R object. Gives the output of native serialization
methods from the model-supplying package, sometimes with additional
classes or attributes that aid portability. This is often
a \link[base:raw]{raw} object.}
\item{situate}{A function. The \code{situate()} function is defined when
\code{\link[=bundle]{bundle()}} is called, though is a loose analogue of an \code{\link[=unbundle]{unbundle()}} S3
method for that object. Since the function is defined on \code{\link[=bundle]{bundle()}}, it
has access to references and dependency information that can
be saved alongside the \code{object} component. Calling \code{\link[=unbundle]{unbundle()}} on a
bundled object \code{x} calls \code{x$situate(x$object)}, returning the
unserialized version of \code{object}. \code{situate()} will also restore needed
references, such as server instances and environmental variables.}

Bundles are R objects that represent a "standalone" version of their
analogous model object. Thus, bundles are ready for saving to file---saving
with \code{\link[base:readRDS]{base::saveRDS()}} is our recommended serialization strategy for bundles,
unless documented otherwise for a specific method.

To restore the inputted model object \code{x} in a new environment, load its
bundle with \code{\link[base:readRDS]{base::readRDS()}} and run \code{\link[=unbundle]{unbundle()}} on it. The output
of \code{\link[=unbundle]{unbundle()}} is a model object that is ready to \code{\link[=predict]{predict()}} on new data,
and other restored functionality (like plotting or summarizing) is supported
as a side effect only.

See \code{vignette("bundle")} for more information on bundling and its motivation.
}
\description{
Bundling a model prepares it to be saved to file and later
restored for prediction in a new R session. See the 'Value' section for
more information on bundles and their usage.
}
\examples{
\dontshow{if (rlang::is_installed("lightgbm")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# fit model and bundle ------------------------------------------------
library(lightgbm)

set.seed(1)

cars_train <-
  lgb.Dataset(
    data = as.matrix(mtcars[1:25, 2:ncol(mtcars)]),
    label = mtcars[1:25, 1],
    params = list(feature_pre_filter = "false")
  )

cars_test <- as.matrix(mtcars[26:32, 2:ncol(mtcars)])

lgb_fit <-
  lgb.train(
    params = list(
      max_depth = 3,
      min_data_in_leaf = 5,
      objective = "regression"
    ),
    data = cars_train,
    nrounds = 5,
    verbose = -1
  )

lgb_bundle <- bundle(lgb_fit)

# then, after saveRDS + readRDS or passing to a new session ----------
lgb_unbundled <- unbundle(lgb_bundle)

lgb_unbundled_preds <- predict(lgb_unbundled, cars_test)
\dontshow{\}) # examplesIf}
}
\seealso{
This method makes use of the \code{save_model_to_string()} R6 methods
of \code{lgb.Booster} objects as well as \code{\link[lightgbm:lgb.load]{lightgbm::lgb.load()}}.

Other bundlers: 
\code{\link{bundle.H2OAutoML}()},
\code{\link{bundle.Learner}()},
\code{\link{bundle.keras.engine.training.Model}()},
\code{\link{bundle.luz_module_fitted}()},
\code{\link{bundle.model_fit}()},
\code{\link{bundle.recipe}()},
\code{\link{bundle.step_umap}()},
\code{\link{bundle.train}()},
\code{\link{bundle.workflow}()},
\code{\link{bundle.xgb.Booster}()},
\code{\link{bundle}()}
}
\concept{bundlers}
