% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_bundler}
\alias{has_bundler}
\title{Check whether an object has a bundling method}
\usage{
has_bundler(x)
}
\arguments{
\item{x}{A model object to bundle.}
}
\value{
A logical.
}
\description{
Given a model object, this function will return whether the object
will dispatch to a method other than \link{bundle.default} (the identity
function).
}
\details{
Note that a return value of \code{FALSE} does not necessarily mean that
the object \code{x} cannot be saved and re-loaded in a new session---many model
objects, like \code{\link[stats:lm]{stats::lm()}} and \code{\link[stats:glm]{stats::glm()}} output, can be effectively
saved and re-loaded in a new session without any bundling.
}
\seealso{
\code{\link[=bundle]{bundle()}}, \code{\link[=unbundle]{unbundle()}}
}
