% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bundle_constr}
\alias{bundle_constr}
\alias{situate_constr}
\alias{swap_element}
\title{Internal Functions}
\usage{
bundle_constr(object, situate, desc_class)

situate_constr(fn)

swap_element(x, ...)
}
\value{
The two \verb{_constr()} functions are constructors that return a bundle
and a situater, respectively. \code{swap_element()} returns \code{x} after swapping
out the specified element.
}
\description{
These functions are not user-facing and are only exported for developer
extensions.
}
\keyword{internal}
